+++ ClientApp/css/CardList.css
@@ -8,13 +8,14 @@
     height: 40px !important;
     border-bottom: 0 !important;
     font-weight: bold !important;
-    padding: 0.5em 1.25em 0.5em 0.5em !important;
+    padding: 0.5em !important;
     background-color: RGB(200, 200, 200) !important;
 }
 
 .list-name {
     background-color: transparent !important;
-    width: 100% !important;
+    width: calc(100% - 30px - 0.5em) !important;
+    margin-right: 0.5em !important;
     border: 0 !important;
     height: 30px !important;
     font-size: 20px !important;
@@ -44,4 +45,11 @@
     justify-content: center !important;
     padding: 0.5em !important;
     background-color: RGB(200, 200, 200) !important;
+}
+
+.delete-list--icon {
+    width: 30px !important;
+    height: 30px !important;
+    margin: 0 !important;
+    cursor: pointer !important;
 }
