+++ ClientApp/css/CardUpdateModal.css
@@ -1,3 +1,3 @@
-.card-modal {
+.card-update-modal {
     margin-top: 0 !important;
 }
