+++ ClientApp/css/NewCardModal.css
@@ -1,3 +1,7 @@
 .card-modal {
     margin-top: 0 !important;
 }
+
+.new-card--button {
+    margin: 0 !important;
+}
